third_party/WebKit/LayoutTests/fast/forms/time/time-stepup-stepdown-expected.txt

Check stepUp() and stepDown() bahevior for type=time.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Invalid value
PASS stepUp("", null, null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': The form element's current value is not a number..
PASS stepDown("", null, null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': The form element's current value is not a number..
Non-number arguments
PASS stepUp("20:13", null, null, "0") is "20:13"
PASS stepDown("20:13", null, null, "0") is "20:13"
PASS stepUp("20:13", null, null, "foo") is "20:13"
PASS stepDown("20:13", null, null, "foo") is "20:13"
PASS stepUp("20:13", null, null, null) is "20:13"
PASS stepDown("20:13", null, null, null) is "20:13"
Normal cases
PASS stepUp("20:13", null, null) is "20:14"
PASS stepDown("20:13", null, null) is "20:12"
PASS stepUp("20:13", null, null, 10) is "20:23"
PASS stepDown("20:13", null, null, 11) is "20:02"
PASS stepUp("20:13", "4", null, 2) is "20:13:08"
PASS stepDown("20:13", "4", null, 3) is "20:12:48"
Step=any
PASS stepUp("20:13", "any", null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element does not have an allowed value step..
PASS stepDown("20:13", "any", null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element does not have an allowed value step..
Overflow/underflow
PASS stepUp("20:13", "3.40282346e+38", null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': The form element's new value (3.40282346e+41) would be higher than the maximum (86399999), and snapping to the maximum would exceed the amount of acceptible error..
PASS stepDown("20:13", "3.40282346e+38", null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': The form element's new value (-3.40282346e+41) would be lower than the minimum (0), and snapping to the minimum would exceed the amount of acceptible error..
PASS stepUp("20:13", "1", "20:13") threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': The form element's new value (72781000) would be higher than the maximum (72780000), and snapping to the maximum would exceed the amount of acceptible error..
PASS stepDown("20:13", "1", "20:13") threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': The form element's new value (72779000) would be lower than the minimum (72780000), and snapping to the minimum would exceed the amount of acceptible error..
PASS stepUp("23:59", null, null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': The form element's new value (86400000) would be higher than the maximum (86399999), and snapping to the maximum would exceed the amount of acceptible error..
PASS stepDown("00:00", null, null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': The form element's new value (-60000) would be lower than the minimum (0), and snapping to the minimum would exceed the amount of acceptible error..

PASS successfullyParsed is true

TEST COMPLETE
